refactor: utils/ 유틸함수와 서비스에서 사용하는 서비스메서드를 분리
