+++ src/js/index.js
@@ -1,4 +1,4 @@
 import {LottoApp} from './components/LottoApp.js';
-import {$} from './utils/document.js';
+import {$} from './utils/element.js';
 
 LottoApp($('#app'));
