+++ src/js/utils/math.js
@@ -1,0 +1,3 @@
+export function getRandomNumber(min, max) {
+    return Math.floor(Math.random() * (max - min + 1)) + min;
+}
